pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "irst-library"
version = "1.0.0"
description = "Advanced Infrared Small Target Detection & Segmentation Research Library"
authors = [
    {name = "IRST Research Team", email = "contact@irst-lib.org"}
]
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["infrared", "small-target-detection", "computer-vision", "deep-learning", "segmentation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]
requires-python = ">=3.8"
dependencies = [
    "torch>=1.12.0",
    "torchvision>=0.13.0",
    "numpy>=1.21.0",
    "opencv-python>=4.6.0",
    "Pillow>=9.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "tqdm>=4.64.0",
    "tensorboard>=2.9.0",
    "PyYAML>=6.0",
    "scikit-learn>=1.1.0",
    "scipy>=1.9.0",
    "albumentations>=1.3.0",
    "timm>=0.6.0",
    "wandb>=0.13.0",
    "hydra-core>=1.2.0",
    "omegaconf>=2.2.0",
    "rich>=12.0.0",
    "typer>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
    "pre-commit>=2.20.0",
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=0.18.0",
    "sphinx-autoapi>=2.0.0",
]
benchmark = [
    "pandas>=1.5.0",
    "plotly>=5.11.0",
    "streamlit>=1.15.0",
]

[project.urls]
Homepage = "https://github.com/sachin-deshik-10/irst-library"
Documentation = "https://irst-library.readthedocs.io"
Repository = "https://github.com/sachin-deshik-10/irst-library"
"Bug Tracker" = "https://github.com/sachin-deshik-10/irst-library/issues"

[project.scripts]
irst-train = "irst_library.cli:train"
irst-eval = "irst_library.cli:evaluate"
irst-benchmark = "irst_library.cli:benchmark"
irst-demo = "irst_library.cli:demo"

[tool.setuptools.packages.find]
where = ["."]
include = ["irst_library*"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["irst_library"]
omit = ["*/tests/*", "*/examples/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
